Moved title and caption tests to separate file
